+++ src/pages/Mypage/index.jsx
@@ -1,7 +1,7 @@
-import React, { useEffect, useState } from "react"
-import BaseLayout from "../../components/BaseLayout"
-import { useNavigate } from "react-router-dom"
-import AddCareerPage from "./AddCareer/AddCareer"
+import React, { useState } from "react";
+import BaseLayout from "../../components/BaseLayout";
+import { useNavigate } from "react-router-dom";
+import AddCareerPage from "./AddCareer/AddCareer";
 import {
   Wrapper,
   ProfileInfo,
@@ -10,91 +10,34 @@ import {
   InfoGroup,
   InfoSubText,
   FeedbackBtn,
-  CategoryWrapper,
-  CategoryIcon,
-  CategoryName,
-  CategoryGroup,
-  ItemDate,
-  ItemGroup,
-  ItemName,
   ContentWrapper,
   StyledButton,
   NaviWrapper,
   StyledPlus,
-} from "./styled"
-import { ModalOverlay, ModalContent } from "../Calendar/styled"
-import { ReactComponent as Share } from "../../svgs/share.svg"
-import { ReactComponent as Plus } from "../../svgs/plus.svg"
-import { ReactComponent as Setting } from "../../svgs/Settings.svg"
-import { Highlight } from "../../components/Typo"
-import { reqGetFeedback } from "../../apis/feedback"
-import userAtom from "../../store/atoms/user"
-import { useRecoilState } from "recoil"
-import { resumeData } from "../../constants/data"
-import { reqGetUser } from "../../apis/user"
+} from "./styled";
+import { ModalOverlay, ModalContent } from "../Calendar/styled";
+import { ReactComponent as Share } from "../../svgs/share.svg";
+import { ReactComponent as Plus } from "../../svgs/plus.svg";
+import { ReactComponent as Setting } from "../../svgs/Settings.svg";
+import { Highlight } from "../../components/Typo";
+import CategoryItem from "./CategoryItem";
+import { useGroupedData, useFeedback } from "../../hooks/useMypage";
 
 export default function Mypage() {
-  const navigate = useNavigate()
-  const [isModalOpen, setIsModalOpen] = useState(false)
-  const [isFeedbackModalOpen, setIsFeedbackModalOpen] = useState(false)
-  const [feedbackData, setFeedbackData] = useState("")
-  const [userData, setUserData] = useRecoilState(userAtom)
-
-  // 카테고리 이름과 아이콘 매핑
-  const categoryMapping = {
-    academy_list: { name: "학력", icon: "/categories/category1.png" },
-    career_list: { name: "경력", icon: "/categories/category2.png" },
-    qualification_list: { name: "자격·어학·수상", icon: "/categories/category3.png" },
-    experience_list: { name: "경험·활동·교육", icon: "/categories/category4.png" },
-    etc_list: { name: "기타", icon: "/categories/category5.png" },
-  }
-
-  // groupedData 생성
-  const groupedData = Object.keys(resumeData).map((key) => ({
-    id: key,
-    name: categoryMapping[key]?.name || "UNKNOWN",
-    icon: categoryMapping[key]?.icon || "/default_icon.png",
-    items: resumeData[key].map((item) => ({
-      startDate: item.start_date,
-      endDate: item.end_date,
-      itemName: item.title,
-      detail: item.content,
-    })),
-  }))
-
-  useEffect(() => {
-    const fetchUser = async () => {
-      const res = await reqGetUser()
-      if (res) {
-        setUserData({
-          nickname: res.data.nickname,
-          img: res.data.image_url,
-        })
-      }
-    }
-    fetchUser()
-  }, [])
-
-  const handleAddCareer = (newCareer) => {
-    setIsModalOpen(false)
-  }
-
-  const createFeedback = async () => {
-    const res = await reqGetFeedback()
-    if (res) {
-      setFeedbackData(res.data.message) // Assuming `message` contains feedback text
-      setIsFeedbackModalOpen(true)
-    }
-  }
+  const navigate = useNavigate();
+  const groupedData = useGroupedData();
+  const { feedbackData, isFeedbackModalOpen, setIsFeedbackModalOpen, createFeedback } =
+    useFeedback();
+  const [isModalOpen, setIsModalOpen] = useState(false);
 
   return (
     <BaseLayout>
       <Wrapper>
         <ProfileInfo>
-          <ProfileImg src={userData.img} />
+          <ProfileImg src="/default_profile.png" />
           <InfoGroup>
             <InfoText>
-              <Highlight>{userData.nickname}</Highlight>님의 이력
+              <Highlight>사용자</Highlight>님의 이력
             </InfoText>
             <InfoSubText>나의 이력을 추가하고 관리할 수 있어요</InfoSubText>
           </InfoGroup>
@@ -102,31 +45,7 @@ export default function Mypage() {
         <FeedbackBtn onClick={createFeedback}>AI 피드백 받기</FeedbackBtn>
         <ContentWrapper>
           {groupedData.map((category) => (
-            <CategoryWrapper key={category.id}>
-              <CategoryGroup>
-                <CategoryIcon src={category.icon}></CategoryIcon>
-                <CategoryName>{category.name}</CategoryName>
-              </CategoryGroup>
-              {category.items.map((item, index) => (
-                <ItemGroup key={index}>
-                  <ItemDate>
-                    {item.startDate} ~ {item.endDate}
-                  </ItemDate>
-                  <div style={{ display: "flex", flexDirection: "column" }}>
-                    <ItemName>{item.itemName}</ItemName>
-                    <ItemName style={{ fontSize: "12px", color: "#C3C3C3" }}>
-                      {item.detail}
-                    </ItemName>
-                  </div>
-                  <img
-                    src="/optionicon.png"
-                    width="2px"
-                    height="10px"
-                    style={{ marginBottom: "15px" }}
-                  />
-                </ItemGroup>
-              ))}
-            </CategoryWrapper>
+            <CategoryItem key={category.id} category={category}/>
           ))}
           <NaviWrapper>
             <StyledButton>
@@ -135,7 +54,7 @@ export default function Mypage() {
             <StyledButton>
               <Setting
                 onClick={() => {
-                  navigate("/setting")
+                  navigate("/setting");
                 }}
               />
             </StyledButton>
@@ -145,7 +64,7 @@ export default function Mypage() {
           </StyledPlus>
         </ContentWrapper>
         {isModalOpen && (
-          <AddCareerPage setIsModalOpen={setIsModalOpen} onAddCareer={handleAddCareer} />
+          <AddCareerPage setIsModalOpen={setIsModalOpen} />
         )}
         {isFeedbackModalOpen && (
           <ModalOverlay>
@@ -158,5 +77,5 @@ export default function Mypage() {
         )}
       </Wrapper>
     </BaseLayout>
-  )
+  );
 }
